refactor(core): move FileSystemWatcher into the box-tool module
Relocates the FileSystemWatcher module from the top-level `modules` directory to be a sub-module within the `box-tool`.

This change improves code organization by grouping the file watching functionality directly with the `box-tool` feature that primarily utilizes it. All relevant import paths have been updated.
