+++ apps/core-app/src/main/modules/box-tool/addon/apps/app-provider.ts
@@ -14,7 +14,7 @@ import { shell } from 'electron'
 import searchEngineCore from '../../search-engine/search-core'
 import path from 'path'
 import fs from 'fs'
-import FileSystemWatcher from '../../../file-system-watcher'
+import FileSystemWatcher from '../../file-system-watcher'
 import { createDbUtils } from '../../../../db/utils'
 import { files as filesSchema, fileExtensions, keywordMappings } from '../../../../db/schema'
 import { and, eq, inArray, or, sql } from 'drizzle-orm'
