+++ apps/core-app/src/main/modules/box-tool/file-system-watcher/index.ts
